+++ app/src/main/java/org/simple/clinic/patient/PatientSearchResult.kt
@@ -1,6 +1,7 @@
 package org.simple.clinic.patient
 
 import android.os.Parcelable
+import androidx.paging.PagingSource
 import androidx.room.Dao
 import androidx.room.DatabaseView
 import androidx.room.Embedded
@@ -115,18 +116,13 @@ data class PatientSearchResult(
   interface RoomDao {
 
     @Query("""
-        SELECT * FROM (
-            SELECT searchResult.*, 1 priority FROM 
-            PatientSearchResult searchResult
-            LEFT JOIN Patient P ON P.uuid = searchResult.uuid
-            WHERE P.status = :status AND P.deletedAt IS NULL AND P.assignedFacilityId = :facilityUuid
-            )
-        ORDER BY priority DESC, fullName COLLATE NOCASE
+      SELECT searchResult.*, 1 priority FROM 
+      PatientSearchResult searchResult
+      LEFT JOIN Patient P ON P.uuid = searchResult.uuid
+      WHERE P.status = :status AND P.deletedAt IS NULL AND P.assignedFacilityId = :facilityId
+      ORDER BY fullName COLLATE NOCASE
     """)
-    fun searchInFacilityAndSortByName(
-        facilityUuid: UUID,
-        status: PatientStatus
-    ): List<PatientSearchResult>
+    fun allPatientsInFacility(facilityId: UUID, status: PatientStatus): PagingSource<Int, PatientSearchResult>
 
     @Query("""
         SELECT