+++ app/src/test/java/org/simple/clinic/instantsearch/InstantSearchResultsItemTypeTest.kt
@@ -3,201 +3,159 @@ package org.simple.clinic.instantsearch
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
 import org.simple.clinic.TestData
-import org.simple.clinic.patient.PatientSearchResult
-import org.simple.clinic.widgets.PatientSearchResultItemView
+import org.simple.clinic.instantsearch.InstantSearchResultsItemType.AssignedFacilityHeader
+import org.simple.clinic.instantsearch.InstantSearchResultsItemType.NearbyFacilitiesHeader
+import org.simple.clinic.instantsearch.InstantSearchResultsItemType.SearchResult
 import java.util.UUID
 
 class InstantSearchResultsItemTypeTest {
 
-  private val currentFacilityUuid = UUID.fromString("69cf85c8-6788-4071-b985-0536ae606b70")
-  private val otherFacilityUuid = UUID.fromString("28ed7dca-828e-4178-942f-fd6937f37155")
-  private val currentFacility = TestData.facility(currentFacilityUuid)
-  private val searchQueryReceived = "A"
-  private val searchQueryEmpty = ""
+  private val currentFacility = TestData.facility(
+      uuid = UUID.fromString("7174598a-434f-4043-9140-8c23046d4ff6")
+  )
+  private val otherFacility = TestData.facility(
+      uuid = UUID.fromString("c7328023-bcd4-4862-b16f-5404df3fd051")
+  )
 
   @Test
-  fun `list items must be generated from the search results`() {
+  fun assigned_facility_header_must_be_shown_when_the_first_search_result_is_at_current_facility() {
     // given
-    val patientUuid1 = UUID.fromString("1d5f18d9-43f7-4e7f-92d3-a4f641709470")
-    val patientUuid2 = UUID.fromString("139bfac5-1adc-43fa-9406-d1000fb67a88")
-    val patientSearchResult1 = TestData.patientSearchResult(patientUuid1, assignedFacilityId = currentFacilityUuid)
-    val patientSearchResult2 = TestData.patientSearchResult(patientUuid2, assignedFacilityId = otherFacilityUuid)
-    val searchResults = listOf(
-        patientSearchResult1,
-        patientSearchResult2
+    val patientInCurrentFacility = SearchResult.forSearchResult(
+        searchResult = TestData.patientSearchResult(
+            uuid = UUID.fromString("2ea03550-896d-4a3d-a977-5ec9c3f91002"),
+            assignedFacilityId = currentFacility.uuid
+        ),
+        currentFacilityId = currentFacility.uuid,
+        searchQuery = null
     )
 
     // when
-    val listItems = InstantSearchResultsItemType.from(
-        patientSearchResults = searchResults,
-        currentFacility = currentFacility,
-        searchQuery = searchQueryReceived
+    val header = InstantSearchResultsItemType.insertHeaders(
+        beforeSearchResult = null,
+        afterSearchResult = patientInCurrentFacility,
+        currentFacility = currentFacility
     )
 
     // then
-    val expected = listOf(
-        InstantSearchResultsItemType.AssignedFacilityHeader(facilityName = currentFacility.name),
-        InstantSearchResultsItemType.SearchResult(
-            searchResultViewModel = PatientSearchResultItemView.PatientSearchResultViewModel(
-                uuid = patientSearchResult1.uuid,
-                fullName = patientSearchResult1.fullName,
-                gender = patientSearchResult1.gender,
-                age = patientSearchResult1.age,
-                dateOfBirth = patientSearchResult1.dateOfBirth,
-                address = patientSearchResult1.address,
-                phoneNumber = patientSearchResult1.phoneNumber,
-                lastSeen = patientSearchResult1.lastSeen,
-                identifier = patientSearchResult1.identifier
-            ),
-            currentFacilityId = currentFacilityUuid,
-            searchQuery = searchQueryReceived
-        ),
-        InstantSearchResultsItemType.NearbyFacilitiesHeader,
-        InstantSearchResultsItemType.SearchResult(
-            searchResultViewModel = PatientSearchResultItemView.PatientSearchResultViewModel(
-                uuid = patientSearchResult2.uuid,
-                fullName = patientSearchResult2.fullName,
-                gender = patientSearchResult2.gender,
-                age = patientSearchResult2.age,
-                dateOfBirth = patientSearchResult2.dateOfBirth,
-                address = patientSearchResult2.address,
-                phoneNumber = patientSearchResult2.phoneNumber,
-                lastSeen = patientSearchResult2.lastSeen,
-                identifier = patientSearchResult2.identifier
-            ),
-            currentFacilityId = currentFacilityUuid,
-            searchQuery = searchQueryReceived
-        )
-    )
-    assertThat(listItems).isEqualTo(expected)
+    assertThat(header).isEqualTo(AssignedFacilityHeader(currentFacility.name))
   }
 
   @Test
-  fun `when there are no search results in other facilities, the other results header must not be generated`() {
+  fun nearby_facility_header_must_be_shown_when_the_first_search_results_is_not_at_current_facility() {
     // given
-    val patientUuid1 = UUID.fromString("1d5f18d9-43f7-4e7f-92d3-a4f641709470")
-    val patientUuid2 = UUID.fromString("139bfac5-1adc-43fa-9406-d1000fb67a88")
-    val patientSearchResult1 = TestData.patientSearchResult(patientUuid1, assignedFacilityId = currentFacilityUuid)
-    val patientSearchResult2 = TestData.patientSearchResult(patientUuid2, assignedFacilityId = currentFacilityUuid)
-    val searchResults = listOf(
-        patientSearchResult1,
-        patientSearchResult2
+    val patientInOtherFacility = SearchResult.forSearchResult(
+        searchResult = TestData.patientSearchResult(
+            uuid = UUID.fromString("2ea03550-896d-4a3d-a977-5ec9c3f91002"),
+            assignedFacilityId = otherFacility.uuid
+        ),
+        currentFacilityId = currentFacility.uuid,
+        searchQuery = null
     )
 
     // when
-    val listItems = InstantSearchResultsItemType.from(
-        patientSearchResults = searchResults,
-        currentFacility = currentFacility,
-        searchQuery = searchQueryEmpty
+    val header = InstantSearchResultsItemType.insertHeaders(
+        beforeSearchResult = null,
+        afterSearchResult = patientInOtherFacility,
+        currentFacility = currentFacility
     )
 
     // then
-    val expected = listOf(
-        InstantSearchResultsItemType.AssignedFacilityHeader(facilityName = currentFacility.name),
-        InstantSearchResultsItemType.SearchResult(
-            searchResultViewModel = PatientSearchResultItemView.PatientSearchResultViewModel(
-                uuid = patientSearchResult1.uuid,
-                fullName = patientSearchResult1.fullName,
-                gender = patientSearchResult1.gender,
-                age = patientSearchResult1.age,
-                dateOfBirth = patientSearchResult1.dateOfBirth,
-                address = patientSearchResult1.address,
-                phoneNumber = patientSearchResult1.phoneNumber,
-                lastSeen = patientSearchResult1.lastSeen,
-                identifier = patientSearchResult1.identifier
-            ),
-            currentFacilityId = currentFacilityUuid,
-            searchQuery = searchQueryEmpty
-        ),
-        InstantSearchResultsItemType.SearchResult(
-            searchResultViewModel = PatientSearchResultItemView.PatientSearchResultViewModel(
-                uuid = patientSearchResult2.uuid,
-                fullName = patientSearchResult2.fullName,
-                gender = patientSearchResult2.gender,
-                age = patientSearchResult2.age,
-                dateOfBirth = patientSearchResult2.dateOfBirth,
-                address = patientSearchResult2.address,
-                phoneNumber = patientSearchResult2.phoneNumber,
-                lastSeen = patientSearchResult2.lastSeen,
-                identifier = patientSearchResult2.identifier
-            ),
-            currentFacilityId = currentFacilityUuid,
-            searchQuery = searchQueryEmpty
-        )
-    )
-    assertThat(listItems).isEqualTo(expected)
+    assertThat(header).isEqualTo(NearbyFacilitiesHeader)
   }
 
   @Test
-  fun `when there are no search results in the current facility, then the nearby facilities should be shown`() {
+  fun nearby_facility_header_must_be_shown_between_search_results_at_current_facility_and_not_at_current_facility() {
     // given
-    val patientUuid1 = UUID.fromString("1d5f18d9-43f7-4e7f-92d3-a4f641709470")
-    val patientUuid2 = UUID.fromString("139bfac5-1adc-43fa-9406-d1000fb67a88")
-    val patientSearchResult1 = TestData.patientSearchResult(patientUuid1, assignedFacilityId = otherFacilityUuid)
-    val patientSearchResult2 = TestData.patientSearchResult(patientUuid2, assignedFacilityId = otherFacilityUuid)
-    val searchResults = listOf(
-        patientSearchResult1,
-        patientSearchResult2
+    val patientInCurrentFacility = SearchResult.forSearchResult(
+        searchResult = TestData.patientSearchResult(
+            uuid = UUID.fromString("2ea03550-896d-4a3d-a977-5ec9c3f91002"),
+            assignedFacilityId = currentFacility.uuid
+        ),
+        currentFacilityId = currentFacility.uuid,
+        searchQuery = null
+    )
+
+    val patientInOtherFacility = SearchResult.forSearchResult(
+        searchResult = TestData.patientSearchResult(
+            uuid = UUID.fromString("39d44632-3183-4755-ae10-966651447197"),
+            assignedFacilityId = otherFacility.uuid
+        ),
+        currentFacilityId = currentFacility.uuid,
+        searchQuery = null
     )
 
     // when
-    val listItems = InstantSearchResultsItemType.from(
-        patientSearchResults = searchResults,
-        currentFacility = currentFacility,
-        searchQuery = searchQueryEmpty
+    val header = InstantSearchResultsItemType.insertHeaders(
+        beforeSearchResult = patientInCurrentFacility,
+        afterSearchResult = patientInOtherFacility,
+        currentFacility = currentFacility
     )
 
     // then
-    val expected = listOf(
-        InstantSearchResultsItemType.NearbyFacilitiesHeader,
-        InstantSearchResultsItemType.SearchResult(
-            searchResultViewModel = PatientSearchResultItemView.PatientSearchResultViewModel(
-                uuid = patientSearchResult1.uuid,
-                fullName = patientSearchResult1.fullName,
-                gender = patientSearchResult1.gender,
-                age = patientSearchResult1.age,
-                dateOfBirth = patientSearchResult1.dateOfBirth,
-                address = patientSearchResult1.address,
-                phoneNumber = patientSearchResult1.phoneNumber,
-                lastSeen = patientSearchResult1.lastSeen,
-                identifier = patientSearchResult1.identifier
-            ),
-            currentFacilityId = currentFacilityUuid,
-            searchQuery = searchQueryEmpty
-
-        ),
-        InstantSearchResultsItemType.SearchResult(
-            searchResultViewModel = PatientSearchResultItemView.PatientSearchResultViewModel(
-                uuid = patientSearchResult2.uuid,
-                fullName = patientSearchResult2.fullName,
-                gender = patientSearchResult2.gender,
-                age = patientSearchResult2.age,
-                dateOfBirth = patientSearchResult2.dateOfBirth,
-                address = patientSearchResult2.address,
-                phoneNumber = patientSearchResult2.phoneNumber,
-                lastSeen = patientSearchResult2.lastSeen,
-                identifier = patientSearchResult2.identifier
-            ),
-            currentFacilityId = currentFacilityUuid,
-            searchQuery = searchQueryEmpty
-        )
-    )
-    assertThat(listItems).isEqualTo(expected)
+    assertThat(header).isEqualTo(NearbyFacilitiesHeader)
   }
 
   @Test
-  fun `when there are no search results, then no list items must be generated`() {
+  fun no_header_should_be_shown_when_both_search_results_are_at_current_facility() {
     // given
-    val searchResults = emptyList<PatientSearchResult>()
+    val patient1InCurrentFacility = SearchResult.forSearchResult(
+        searchResult = TestData.patientSearchResult(
+            uuid = UUID.fromString("2ea03550-896d-4a3d-a977-5ec9c3f91002"),
+            assignedFacilityId = currentFacility.uuid
+        ),
+        currentFacilityId = currentFacility.uuid,
+        searchQuery = null
+    )
+
+    val patient2InCurrentFacility = SearchResult.forSearchResult(
+        searchResult = TestData.patientSearchResult(
+            uuid = UUID.fromString("39d44632-3183-4755-ae10-966651447197"),
+            assignedFacilityId = currentFacility.uuid
+        ),
+        currentFacilityId = currentFacility.uuid,
+        searchQuery = null
+    )
 
     // when
-    val listItems = InstantSearchResultsItemType.from(
-        patientSearchResults = searchResults,
-        currentFacility = currentFacility,
-        searchQuery = searchQueryEmpty
+    val header = InstantSearchResultsItemType.insertHeaders(
+        beforeSearchResult = patient1InCurrentFacility,
+        afterSearchResult = patient2InCurrentFacility,
+        currentFacility = currentFacility
     )
 
     // then
-    assertThat(listItems).isEmpty()
+    assertThat(header).isNull()
+  }
+
+  @Test
+  fun no_header_should_be_shown_when_both_search_results_are_not_at_current_facility() {
+    // given
+    val patient1InOtherFacility = SearchResult.forSearchResult(
+        searchResult = TestData.patientSearchResult(
+            uuid = UUID.fromString("2ea03550-896d-4a3d-a977-5ec9c3f91002"),
+            assignedFacilityId = otherFacility.uuid
+        ),
+        currentFacilityId = currentFacility.uuid,
+        searchQuery = null
+    )
+
+    val patient2InOtherFacility = SearchResult.forSearchResult(
+        searchResult = TestData.patientSearchResult(
+            uuid = UUID.fromString("39d44632-3183-4755-ae10-966651447197"),
+            assignedFacilityId = otherFacility.uuid
+        ),
+        currentFacilityId = currentFacility.uuid,
+        searchQuery = null
+    )
+
+    // when
+    val header = InstantSearchResultsItemType.insertHeaders(
+        beforeSearchResult = patient1InOtherFacility,
+        afterSearchResult = patient2InOtherFacility,
+        currentFacility = currentFacility
+    )
+
+    // then
+    assertThat(header).isNull()
   }
 }