+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Bump Paging to v3.0.0
 - Use new method of SQL performance profiling (transform generated Room DAOs)
 - Prefill search query in `InstantSearchScreen`
+- [In Progress: 10 June 2021] Paginate search results in `InstantSearchScreen`
 
 ### Changes
 - Updated translations: `hi-IN`, `bn-BD`, `te-IN`, `bn-IN`, `am-ET`, `pa-IN`, `ti-ET`, `kn-IN`, `mr-IN`, `ta-IN`, `so-ET`, `om-ET`